fix failing corner-case of diff
`git diff <old-revision> HEAD -- net.db` was working fine

but

`git diff HEAD <old-revision> -- net.db` was giving an empty diff
even though there were differences between the DBs.

the problem turned out to be the nuances of the args passed by git
to the diff driver (sqlite-diff).

we get the /tmp/ param in a different spot depending on
what kind of diff logic/context git is performing
sometimes it's in $2, sometimes in $5
if we don't choose the /tmp/ file, we just end up diffing
the same file against itself and lose the proper diff.

we are always interested in the /tmp/ file which has the version we're comparing to
depending on which arg has the /tmp/ file, we need to change the order of
the sqldiff args for some reason barely unknown to us